% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-niwaconverter.R
\name{niwaconverter}
\alias{niwaconverter}
\title{Converter to move around files taken from NIWA climate change projetions and crop to an area of interest}
\usage{
niwaconverter(niwadd, newdd, foldname, croparea)
}
\arguments{
\item{niwadd}{Folder directory for raw NIWA inputs including a max temp, mintemp, precip, and solar rad netcdf file}

\item{newdd}{Folder directory for where to save raster}

\item{foldname}{Name for new folder}

\item{croparea}{SpatialPolygon areal extent in wgs84 to crop extent}
}
\value{
Folder will be filled with 4 new rasters
}
\description{
This function models soil erosion ecosystem services using the Guerra et al. approach but using NZUSLE as the model
}
