% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-carrying-capacity.m2.R
\name{nzes.carrying.m2}
\alias{nzes.carrying.m2}
\title{Modelling of stock basic carrying capacity by crown equations}
\usage{
nzes.carrying.m2(
  luc,
  slope,
  solar = NA,
  vpdann = NA,
  vpdjan = NA,
  meant,
  precip,
  latitude
)
}
\arguments{
\item{luc}{Land use capability index. Integer, values 1-8}

\item{slope}{Slope in degrees.}

\item{solar}{Optional. Total solar radiation in W per m2}

\item{vpdann}{Optional. Annual mean vapour pressure deficit}

\item{vpdjan}{Optional. Mean vapour pressure deficit in January}

\item{meant}{Mean annual air temperature}

\item{precip}{Total annual precipitation}

\item{latitude}{Latitude from WGS as a raster}

\item{elev}{Elevation in metres above sea level}

\item{soilt}{Optional. Mean annual soil temperature in degrees C}

\item{aspect180}{Value in degrees from 0 to 180. North is 0, south is 180. East or west are 90. Flat areas with a slope less than 7degrees have aspect 180 =0}
}
\value{
Raster indicating modelled potential basic carrying capacity
}
\description{
This function models stock carrying capacity. Use m1 for model 1, m2 for model 2
}
