% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-ecocrop-monthly.R
\name{nzes.crop.monthly}
\alias{nzes.crop.monthly}
\title{Model suitability for a crop based on climate and other parameters}
\usage{
nzes.crop.monthly(crop, tn, tx, pr, soils)
}
\arguments{
\item{crop}{Scientific name of the crop for searching ECOCROP}

\item{tn}{Stack of min average temperature}

\item{tx}{Stack of max temperature}

\item{pr}{Stack of monthly precipitation}

\item{soils}{Stack of soils with parameters "ph", "salinity", "prd"(depth) from the NZ soil dataset}
}
\value{
Raster indicating suitability (binary) for the crop or not
}
\description{
This function models suitability for crops using the rules in ECOCROP based on monthly variation. Improved version of nzes.crop
}
