% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-ecocrop.R
\name{nzes.crop}
\alias{nzes.crop}
\title{Model suitability for a crop based on climate and other parameters}
\usage{
nzes.crop(crop, tn, tx, pr, soils)
}
\arguments{
\item{crop}{Scientific name of the crop for searching ECOCROP}

\item{tn}{Raster of mean temperature}

\item{tx}{Raster of max temperature}

\item{pr}{Stack of monthly precipitation}

\item{soils}{Stack of soils with parameters "ph", "salinity", "prd"(depth) from the NZ soil dataset}
}
\value{
Raster indicating suitability (binary) for the crop or not
}
\description{
This function models suitability for crops using the rules in ECOCROP
}
