% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-finnovation.R
\name{nzes.farminnovation}
\alias{nzes.farminnovation}
\title{Farmer innovation from Brown et al 2016 LUP}
\usage{
nzes.farminnovation(agribase, region, srdmlookup, luc)
}
\arguments{
\item{agribase}{Spatial polygons dataframe from agribase including parameters}

\item{region}{Vector aligned with agribase including names of regions for all farms}

\item{srdmlookup}{Lookup table provided in package}

\item{luc}{Land cover map (LUCAS) rasterised as an integer version e.g. 71000, etc.}
}
\value{
Raster layer with values between 0 and 10 indicating innovation index
}
\description{
This function models suitability for 6 crops under irrigated and non irrigated situations
}
