% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-wildfire-fuels.R
\name{nzes.fuel}
\alias{nzes.fuel}
\title{Wildfire fuel map based on land cover and veg height and BUI data}
\usage{
nzes.fuel(BUI, lcm, vegheight, fueltable, fuelconverter)
}
\arguments{
\item{BUI}{Raster of BUI from nzes.wildfire}

\item{lcm}{Land/ water cover map, including vegetation categories.}

\item{vegheight}{Raster of vegetation height in m}

\item{fueltable}{Lookup table of fuel loads}

\item{fuelconverter}{Lookup table to reclasify lcdb to fuel types in fueltable}
}
\value{
Raster indicating fuel load
}
\description{
This function models fuel based on landcover, vegheight, and BUI (for forest)
}
