% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-lcdb-mapper.R
\name{nzes.lcdb.mapper}
\alias{nzes.lcdb.mapper}
\title{Plot a nice map based on an lcdb class raster}
\usage{
nzes.lcdb.mapper(lcm, lcdblookup)
}
\arguments{
\item{lcm}{Raster using lcdb classes}

\item{lcdblookup}{Lookup table of lcdb classes to colours and names}
}
\value{
Plots a tmap object
}
\description{
This function is a helper to get land cover maps for different time points (fixed)
}
