% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-lookup.R
\name{nzes.lookup}
\alias{nzes.lookup}
\title{General function for lookup table type ecosystem service model}
\usage{
nzes.lookup(lc, lt)
}
\arguments{
\item{lc}{Land cover or other integer raster representing categories}

\item{lt}{Lookup table with same number of classes. 2 column matrix}
}
\value{
Raster indicating ecosystem service provision
}
\description{
This function assigns values from a lookup table to a raster. It is just a wrapper for raster::reclassify, so better to use that function really...
}
