% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-niwaconverter.R
\name{nzes.niwaconverter}
\alias{nzes.niwaconverter}
\title{Extract key climate variables for hydrology, erosion, and stock unit carrying capacity models}
\usage{
nzes.niwaconverter(niwadd, reprojraster)
}
\arguments{
\item{niwadd}{Folder for source NIWA netcdfs}

\item{reprojraster}{Raster of resolution needed for modelling, to force NIWA layers into}
}
\value{
Named raster stack with 4 layers.
}
\description{
This function is a helper to get useful layers from the NIWA archives
}
