% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-pfrcrop.R
\name{nzes.pfrcrop}
\alias{nzes.pfrcrop}
\title{Crop suitability for 6 key crops based on rules developed by PFR}
\usage{
nzes.pfrcrop(niwadd, region, soils, croprules)
}
\arguments{
\item{niwadd}{Folder for source NIWA netcdfs}

\item{region}{Cropping region as spatial polygons in wgs84}

\item{soils}{Stack of soil parameters including NAMED LAYERS; drainclass, ph, prd, slope, paw, salinity, clay,}

\item{croprules}{Table containing the crop rules. This is very specific and designed to make this function unusable without permission to use the dataset from PFR. must have 6 columns with one named "sc" for suitability score}
}
\value{
Named raster stack with 10 layers. Not all crops are water limited, so they do not show irrigated/non difference
}
\description{
This function models suitability for 6 crops under irrigated and non irrigated situations
}
