% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-recreationGrigulis.R
\name{nzes.recreationGrigulis}
\alias{nzes.recreationGrigulis}
\title{Model an index of recreation based roughly on Karl's approach}
\usage{
nzes.recreationGrigulis(
  lcm,
  dem,
  protect,
  access,
  huts,
  touristspot,
  agfactor,
  aesrecl,
  viewrecl
)
}
\arguments{
\item{lcm}{Raster of land cover classes as integer}

\item{dem}{Digital elevation model raster}

\item{protect}{Raster of protected areas as binary}

\item{access}{Raster of road and track network}

\item{huts}{Raster of hut locations or other minor attractants and their impact score}

\item{touristspot}{Raster of major tourism attractants and their impact score}

\item{agfactor}{Value for aggregation of raster in calculation recommend 50}

\item{aesrecl}{Matrix for reclassification of aesthetic values on lcdb}

\item{viewrecl}{Matrix for reclassification of view distance on lcdb}
}
\value{
Raster layer of modelled recreational opportunity
}
\description{
This function models an indicator of recreational opportunity spectrum
}
