% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-vegclimates.R
\name{nzes.vegclimates}
\alias{nzes.vegclimates}
\title{Crop suitability for 6 key crops based on rules developed by PFR}
\usage{
nzes.vegclimates(niwadd, reprojraster)
}
\arguments{
\item{niwadd}{Folder for source NIWA netcdfs}

\item{reprojraster}{Raster of resolution needed for modelling, to force NIWA layers into}
}
\value{
Named raster stack with 5 layers.
}
\description{
This function models suitability for 6 crops under irrigated and non irrigated situations
}
