% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-wildfire.R
\name{nzes.wildfire}
\alias{nzes.wildfire}
\title{Wildfire risk indices based on weather}
\usage{
nzes.wildfire(niwadd, region)
}
\arguments{
\item{niwadd}{Folder for source NIWA netcdfs}

\item{region}{Cropping region as spatial polygons in wgs84}
}
\value{
Named raster stack with 3 layers. Gives 3 fire indices - BUI, DC, DMC
}
\description{
This function models wildfire risk based on climate
}
