% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCS-soil-ero.R
\name{soilerosion.model}
\alias{soilerosion.model}
\title{Modelling of soil erosion protection by ecosystems}
\usage{
soilerosion.model(lcm, R, K, LS, ndvi)
}
\arguments{
\item{lcm}{Land/ water cover map}

\item{R}{Rainfall erosivity (MJ/  mm / year). Can be either a raster or single value}

\item{K}{Soil erodibility factor. Can be either a raster with K values for different land cover types or a single value}

\item{LS}{Topographic factors. Must be a raster. Creating an LS raster is straightforward but computationally expensive in R, so a function is not given.}

\item{ndvi}{Normalised difference vegetation index}
}
\value{
List of three rasters - the ecosystem service ES, structural impact SI, and absolute soil erosion A
}
\description{
This function models soil erosion
}
