% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-curvenumber.R
\name{nzes.curvenumber}
\alias{nzes.curvenumber}
\title{Model an index of recreation based roughly on Karl's approach}
\usage{
nzes.curvenumber(curven, precip)
}
\arguments{
\item{curven}{Raster of curve number values between 0 and 100}

\item{precip}{Raster of precipitation}
}
\value{
Raster indicating Q (stormwater runoff in mm)
}
\description{
This function models an indicator of recreational opportunity spectrum
}
