% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-finnovation.R
\name{nzes.farminnovation}
\alias{nzes.farminnovation}
\title{Farmer innovation from Brown et al 2016 LUP}
\usage{
nzes.farminnovation(agribase, region, srdmlookup)
}
\arguments{
\item{agribase}{Spatial polygons dataframe from agribase including parameters}

\item{region}{Vector aligned with agribase including names of regions for all farms}

\item{srdmlookup}{Lookup table provided in package}

\item{lcm}{Land cover map rasterised}
}
\value{
Raster layer
}
\description{
This function models suitability for 6 crops under irrigated and non irrigated situations
}
