% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-pfrcrop.R
\name{nzes.pfrcrop}
\alias{nzes.pfrcrop}
\title{Crop suitability for 6 key crops based on rules developed by PFR}
\usage{
nzes.pfrcrop(tmax, tmin, precip, soils, croprules)
}
\arguments{
\item{tmax}{List ot stacked rasters of max daily temperature. Each stack 1 year}

\item{tmin}{List ot stacked rasters of min daily temperature. Each stack 1 year}

\item{precip}{List ot stacked rasters of daily precip. Each stack 1 year}

\item{soils}{Stack of soil parameters including NAMED LAYERS; drainclass, ph, prd, slope, paw, salinity, clay,}

\item{croprules}{Table containing the crop rules. This is very specific and designed to make this function unusable without permission to use the dataset from PFR}
}
\value{
Named raster stack with 10 layers. Not all crops are water limited, so they do not show irrigated/non difference
}
\description{
This function models suitability for 6 crops under irrigated and non irrigated situations
}
