% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZES-pollination.R
\name{nzes.pollination}
\alias{nzes.pollination}
\title{Model pollination and honey bee potential hive density}
\usage{
nzes.pollination(
  lcm,
  lcclasses,
  pollenlt,
  nectarlt,
  pollendemand,
  pollenrange,
  nectarrange,
  nectardemand = 200
)
}
\arguments{
\item{lcm}{Raster of land cover classes as integer}

\item{lcclasses}{Vector of land cover lookup classes in same order as lookup tables below}

\item{pollenlt}{Lookup table of pollen production kgha-1 per 12 months}

\item{nectarlt}{Lookup table of nectar production kgha-1 per 12 months}

\item{pollendemand}{Vector of pollen demand per month kg}

\item{pollenrange}{Vector of pollen travel distance per month km}

\item{nectarrange}{Vector of nectar travel distance per month km}

\item{nectardemand}{Annual nectar requirement kg}
}
\value{
Raster layer indicating the annual carrying capacity in colonies per pixel
}
\description{
This function models floral resources for honey bee pollinators, following Ausseil et al. 2018 Ecological Applications
}
